fix(Personagem/Batalha): Merge
Resolve conflitos de merge
